move updating SessionInfo with websocket data from App to SessionInfoProvider
